apps/backend/requirements-base.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
flask-marshmallow==0.15.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.12.1

# Serialization and validation
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Authentication and security
python-dotenv==1.0.0
cryptography==41.0.7
PyJWT==2.8.0
werkzeug==3.0.1

# Email
Flask-Mail==0.9.1

# HTTP client for AI service
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1
aiofiles==23.2.1

# Redis for caching - high performance analytics
redis==5.0.1

# File handling
python-magic==0.4.27
mutagen==1.47.0

# Video/Audio URL processing
yt-dlp>=2024.1.0
ffmpeg-python==0.2.0

# Utilities
pytz==2023.3
python-dateutil==2.8.2

# Production server
gunicorn==21.2.0

# API Documentation
flask-restx==1.3.0

# System monitoring for health endpoint
psutil==5.9.6

# WebSocket support for real-time progress updates
flask-socketio==5.3.6
python-socketio==5.11.0
python-engineio==4.9.0
eventlet==0.33.3
gevent==23.9.1
gevent-websocket==0.10.1

# Export functionality
reportlab==4.0.7
python-docx==1.1.0

# Data processing and analytics
numpy==1.26.2
pandas==2.1.4